requirements.txt
geopandas==0.6.2
pandas==0.25.3
base58==2.0.0
matplotlib==3.1.3
numpy==1.18.1
lxml==4.5.1
streamlit==0.68.1
folium==0.12.1.post1
keplergl==0.3.2
geemap==0.9.5
Pillow==8.4.0
leafmap==0.7.0
